Move legacy mappings to utils
