+++ src/sentry/api/serializers/models/user_notifications.py
@@ -3,11 +3,11 @@ from typing import Iterable
 
 from sentry.api.serializers import Serializer
 from sentry.models import NotificationSetting, UserOption
-from sentry.notifications.legacy_mappings import (
+from sentry.notifications.types import FineTuningAPIKey, NotificationScopeType
+from sentry.notifications.utils.legacy_mappings import (
     get_type_from_fine_tuning_key,
     map_notification_settings_to_legacy,
 )
-from sentry.notifications.types import FineTuningAPIKey, NotificationScopeType
 from sentry.types.integrations import ExternalProviders
