+++ src/sentry/notifications/utils/legacy_mappings.py
